完善多表联查

+++ internal/service/comment.go
@@ -1,8 +1,10 @@
 package service
 
 import (
+	"errors"
 	"github.com/fitenne/youthcampus-dousheng/internal/repository"
 	"github.com/fitenne/youthcampus-dousheng/pkg/model"
+	"strconv"
 )
 
 var commentCtl = repository.GetCommentCtl()
@@ -11,7 +13,16 @@ func Publish(videoId int64, comment *model.Comment) error {
 	return commentCtl.Publish(videoId, comment)
 }
 
-func DeleteById(commentId int64) error {
+func DeleteById(userId, commentId int64) error {
+	comment, err := commentCtl.QueryById(commentId)
+	if err != nil {
+		return errors.Unwrap(err)
+	}
+
+	if userId != comment.User.ID {
+		return errors.New("该用户没有权限：" + strconv.Itoa(int(userId)))
+	}
+
 	return commentCtl.DeleteById(commentId)
 }
